# backend/openapi.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "Global Multi-Entity ERP API",
    "version": "0.4.0",
    "description": "API contract for global multi-entity accounting endpoints under /api/v1.",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.global-ledger.com",
      "description": "Production"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/intercompany/pairs": {
      "post": {
        "tags": [
          "Intercompany"
        ],
        "operationId": "upsertIntercompanyPair",
        "summary": "Create or update intercompany pair",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntercompanyPairInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Intercompany pair created or updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntercompanyPairResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/intercompany/entity-flags": {
      "get": {
        "tags": [
          "Intercompany"
        ],
        "operationId": "listIntercompanyEntityFlags",
        "summary": "List legal entity intercompany flags",
        "parameters": [
          {
            "in": "query",
            "name": "tenantId",
            "required": false,
            "description": "Tenant identifier",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "in": "query",
            "name": "legalEntityId",
            "required": false,
            "description": "Legal entity identifier",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Intercompany entity flags",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnyObject"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/intercompany/entity-flags/{legalEntityId}": {
      "patch": {
        "tags": [
          "Intercompany"
        ],
        "operationId": "updateIntercompanyEntityFlags",
        "summary": "Update intercompany flags for legal entity",
        "parameters": [
          {
            "in": "path",
            "name": "legalEntityId",
            "required": true,
            "description": "Legal entity identifier",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnyObject"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Intercompany flags updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnyObject"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/intercompany/reconcile": {
      "post": {
        "tags": [
          "Intercompany"
        ],
        "operationId": "reconcileIntercompany",
        "summary": "Reconcile intercompany balances",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnyObject"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Intercompany reconciliation result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntercompanyReconcileResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "501": {
            "description": "Not implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "ErrorResponse": {
        "description": "Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "AnyObject": {
        "type": "object",
        "additionalProperties": true
      },
...
      "IntercompanyPairResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "id": {
            "type": "integer",
            "nullable": true
          },
          "tenantId": {
            "type": "integer",
            "minimum": 1
          }
        },
        "required": [
          "ok",
          "tenantId"
        ]
      },
      "IntercompanyReconcileResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          }
        },
        "required": [
          "ok",
          "message",
          "items"
        ]
      },
...
      "IntercompanyPairInput": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "integer",
            "minimum": 1
          },
          "fromLegalEntityId": {
            "type": "integer",
            "minimum": 1
          },
          "toLegalEntityId": {
            "type": "integer",
            "minimum": 1
          },
          "receivableAccountId": {
            "type": "integer",
            "minimum": 1,
            "nullable": true
          },
          "payableAccountId": {
            "type": "integer",
            "minimum": 1,
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ]
          }
        },
        "required": [
          "fromLegalEntityId",
          "toLegalEntityId"
        ]
      },
...
    }
  }
}